# YJIT: Decrease IVAR_MAX_DEPTH to 8