I tested 5, 8, and 10 as `GET_IVAR_MAX_DEPTH` and `SET_IVAR_MAX_DEPTH` on SFR. depth=8 showed the best latency, and it had almost the same max code size as depth=5, which is 1.6% smaller than depth=10's.