@maximecb 2023-09-07
Thanks for running this experiment 🙏 

Would just like to see the perf impact of depth 8 vs 10 on headline benchmarks before we merge.